refactor(database): cleanup schema processing and utilities
