+++ modules/database/src/adapters/sequelize-adapter/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './populate';
+export * from './query';
